backend/requirements_dev.txt
asgiref==3.6.0
Django==3.2
djangorestframework==3.11.1 # 3.12.0 # <- на этой версии не работает djoser
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
djoser==2.1.0
flake8
pep8-naming
flake8-broken-line
flake8-return
flake8-isort
Pillow==9.4.0